requirements.txt
streamlit
faiss-cpu
sentence-transformers
python-pptx
python-docx
pypdf
pandas
openai